requirements.txt
Flask-SQLAlchemy==2.5.1
flask-cors
pyjwt
autopep8
Flask-Migrate
pytest
coverage
requests
flask-restx